# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/verify-email": {
      "post": {
        "operationId": "EmailVerificationController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Confirm an email address with the emailed token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verify-email/resend": {
      "post": {
        "description": "Does nothing when the address is already verified.",
        "operationId": "EmailVerificationController_resend",
        "parameters": [],
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationSent"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a new verification email to the signed-in user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/tenants/{tenantId}/customers": {
      "post": {
        "operationId": "EndCustomersController_create",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEndCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndCustomer"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      },
      "get": {
        "operationId": "EndCustomersController_findAll",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "skip",
                    "take"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/EndCustomer"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "skip": {
                      "type": "integer"
                    },
                    "take": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/v1/tenants/{tenantId}/customers/{id}": {
      "get": {
        "operationId": "EndCustomersController_findOne",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndCustomer"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      },
      "patch": {
        "operationId": "EndCustomersController_update",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEndCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndCustomer"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      },
      "delete": {
        "operationId": "EndCustomersController_remove",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndCustomer"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "customers"
        ]
      }
    },
    "/v1/tenants/{tenantId}/customers/{id}/conversations": {
      "get": {
        "description": "The customer view: the same list as `GET …/conversations?customerId=`, newest activity first. A customer of another tenant is a 404 CUSTOMER_NOT_FOUND.",
        "operationId": "EndCustomersController_conversationsOf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "One or more statuses, comma-separated: `active` (the AI answers), `escalated` (waiting for a human, the queue), `human_active` (a human handles it), `resolved`.",
            "schema": {
              "example": "escalated,human_active",
              "type": "string"
            }
          },
          {
            "name": "assignedTo",
            "required": false,
            "in": "query",
            "description": "`me` = the conversations I handle, or the id of a team member.",
            "schema": {
              "example": "me",
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "description": "Only the conversations of this end customer.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "`lastMessageAt`: newest activity first. `escalatedAt`: the queue order, the one that has waited longest first.",
            "schema": {
              "default": "lastMessageAt",
              "type": "string",
              "enum": [
                "lastMessageAt",
                "escalatedAt"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "total",
                    "skip",
                    "take"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Conversation"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "skip": {
                      "type": "integer"
                    },
                    "take": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "CUSTOMER_NOT_FOUND"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Conversations of one customer (every role)",
        "tags": [
          "customers"
        ]
      }
    },
...
    "/v1/i18n/{locale}/{namespace}": {
      "get": {
        "description": "A flat map of stable dot-notation keys to ICU message strings. Keys missing in the requested language are filled in from `en`. Supports `If-None-Match` (304). For API errors use namespace `errors` with the error `code` as key.",
        "operationId": "I18nController_namespace",
        "parameters": [
          {
            "name": "locale",
            "required": true,
            "in": "path",
            "description": "A locale code from GET /v1/i18n/locales.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "namespace",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "common",
                "errors",
                "notifications",
                "widget"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "LOCALE_NOT_FOUND / NAMESPACE_NOT_FOUND"
          }
        },
        "summary": "Translations of one namespace (public, cacheable)",
        "tags": [
          "i18n"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-tenant support platform: backend core",
    "description": "Tenants, staff authentication, invitations and roles, audit log, translations, end customers. Errors always have the shape `{ statusCode, code, message }`; lists are `{ data, total, skip, take }`.",
    "version": "0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VerifyEmailDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The token from the emailed verification link"
          }
        },
        "required": [
          "token"
        ]
      },
      "VerificationSent": {
        "type": "object",
        "properties": {
          "link": {
            "type": "string",
            "x-dev-only": true,
            "description": "DEVELOPMENT ONLY: present only when the server runs with MAIL_MODE=link (refused in production); the verification link."
          }
        }
      },
...
      "ConversationCustomer": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "externalId": {
            "type": "string",
            "nullable": true,
            "example": "web_a1b2c3…",
            "description": "The customer's id in their channel: a phone number, the tenant's own id, or for an anonymous website visitor a SHORTENED `web_` id (the full one is that visitor's secret). Null when the customer record no longer exists."
          },
          "channel": {
            "type": "string",
            "enum": [
              "widget",
              "whatsapp",
              "voice"
            ]
          }
        },
        "required": [
          "name",
          "externalId",
          "channel"
        ]
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "endCustomerId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/ConversationCustomer"
          },
          "channel": {
            "type": "string",
            "enum": [
              "widget",
              "whatsapp",
              "voice"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "escalated",
              "human_active",
              "resolved"
            ],
            "description": "`active` the AI answers; `escalated` waits for a human; `human_active` a human handles it (the AI is silent); `resolved` is over."
          },
          "assignedUserId": {
            "type": "string",
            "nullable": true,
            "description": "The staff member handling it (while `human_active`)."
          },
          "assignedUserName": {
            "type": "string",
            "nullable": true,
            "description": "Their name (or email when they have no name). Null when nobody holds it or the user was deleted."
          },
          "escalatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "escalationReason": {
            "type": "string",
            "enum": [
              "customer_requested",
              "low_confidence",
              "failed_action",
              "sentiment",
              "topic",
              "outside_hours",
              "ai_unavailable",
              "limit_reached",
              "other"
            ],
            "nullable": true,
            "description": "Stable code; translate it in the dashboard."
          },
          "summary": {
            "type": "string",
            "nullable": true,
            "description": "Hand-off note written when it escalated."
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolvedBy": {
            "type": "string",
            "enum": [
              "ai",
              "human",
              "customer",
              "system"
            ],
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "endCustomerId",
          "customer",
          "channel",
          "status",
          "assignedUserId",
          "assignedUserName",
          "escalatedAt",
          "escalationReason",
          "summary",
          "lastMessageAt",
          "resolvedAt",
          "resolvedBy",
          "createdAt"
        ]
      },
...
      "CreateEndCustomerDto": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "example": "web_5f1c...",
            "description": "Unique per tenant (B4)"
          },
          "name": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "example": "en",
            "description": "The customer's language (H7); unset means unknown. A locale code from GET /v1/i18n/locales (BCP-47, for example `en`, `ur`)."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "externalId"
        ]
      },
      "EndCustomer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "externalId": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "locale": {
            "type": "string",
            "nullable": true,
            "example": "ur"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "externalId",
          "createdAt"
        ]
      },
      "UpdateEndCustomerDto": {
        "type": "object",
        "properties": {
          "externalId": {
            "type": "string",
            "example": "web_5f1c...",
            "description": "Unique per tenant (B4)"
          },
          "name": {
            "type": "string"
          },
          "locale": {
            "type": "string",
            "example": "en",
            "description": "The customer's language (H7); unset means unknown. A locale code from GET /v1/i18n/locales (BCP-47, for example `en`, `ur`)."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}